+++ LoopBot/Program.cs
@@ -102,77 +102,23 @@ class Program
             var nftListingDetails = await loopExchangeWebApiService.GetNftListingDetailsAsync(nftFullId);
             var nftTakerListingDetails = await loopExchangeApiService.GetTakerListingDetailsAsync(nftListingDetails.Id);
             var listingPriceDecimal = Utils.ConvertStringToDecimal(nftTakerListingDetails.Erc20TokenAmount);
+
             if (listingPriceDecimal <= priceToBuyDecimal) //NFT has to be under or equal to the price limit
             {
                 Console.WriteLine("Valid Listing found for NFT...Attempting to buy.");
+                //Get  fees
                 var orderFee = await loopringApiService.GetOrderFee(settings.LoopringAccountId, nftDetails.TokenAddress, nftTakerListingDetails.Erc20TokenAmount);
                 var takerOrderFee = await loopExchangeApiService.GetTakerFeesAsync(nftListingDetails.Id, settings.LoopringAccountId, nftDetails.TokenAddress, orderFee.FeeRate.Rate, nftTakerListingDetails.Erc20TokenAmount);
                 var storageId = await loopringApiService.GetNextStorageId(settings.LoopringAccountId, 1);
-                int nftTokenId = nftTakerListingDetails.NftTokenId;
-                string nftData = nftTakerListingDetails.NftTokenData;
 
-                //Creating the nft taker order
-                NftOrder nftTakerOrder = new NftOrder()
-                {
-                    exchange = settings.Exchange,
-                    accountId = settings.LoopringAccountId,
-                    storageId = storageId.orderId,
-                    sellToken = new SellToken
-                    {
-                        tokenId = 1,
-                        amount = nftTakerListingDetails.Erc20TokenAmount
-                    },
-                    buyToken = new BuyToken
-                    {
-                        tokenId = nftTokenId,
-                        nftData = nftData,
-                        amount = "1"
-                    },
-                    allOrNone = false,
-                    fillAmountBOrS = true,
-                    validUntil = DateTimeOffset.Now.AddDays(30).ToUnixTimeSeconds(),
-                    maxFeeBips = orderFee.FeeRate.Rate
-                };
-                int fillAmountBOrSValue2 = 0;
-                if (nftTakerOrder.fillAmountBOrS == true)
-                {
-                    fillAmountBOrSValue2 = 1;
-                }
-
-                BigInteger[] poseidonTakerOrderInputs =
-                {
-                    Utils.ParseHexUnsigned(settings.Exchange),
-                    (BigInteger) nftTakerOrder.storageId,
-                    (BigInteger) nftTakerOrder.accountId,
-                    (BigInteger) nftTakerOrder.sellToken.tokenId,
-                    !String.IsNullOrEmpty(nftTakerOrder.buyToken.nftData) ? Utils.ParseHexUnsigned(nftTakerOrder.buyToken.nftData) : (BigInteger) nftTakerOrder.buyToken.tokenId ,
-                    !String.IsNullOrEmpty(nftTakerOrder.sellToken.amount) ? BigInteger.Parse(nftTakerOrder.sellToken.amount) : (BigInteger) 0,
-                    !String.IsNullOrEmpty(nftTakerOrder.buyToken.amount) ? BigInteger.Parse(nftTakerOrder.buyToken.amount) : (BigInteger) 0,
-                    (BigInteger) nftTakerOrder.validUntil,
-                    (BigInteger) nftTakerOrder.maxFeeBips,
-                    (BigInteger) fillAmountBOrSValue2,
-                    Utils.ParseHexUnsigned("0x0000000000000000000000000000000000000000")
-                };
-
-                //Generate the poseidon hash
-                Poseidon poseidon2 = new Poseidon(12, 6, 53, "poseidon", 5, _securityTarget: 128);
-                BigInteger takerOrderPoseidonHash = poseidon2.CalculatePoseidonHash(poseidonTakerOrderInputs);
-
-                //Generate the signaures
-                Eddsa eddsa2 = new Eddsa(takerOrderPoseidonHash, settings.LoopringPrivateKey);
-                string takerEddsaSignature = eddsa2.Sign();
-                nftTakerOrder.eddsaSignature = takerEddsaSignature;
-                var nftTakerTradeValidateResponse = await loopringApiService.SubmitNftTradeValidateOrder(nftTakerOrder, takerEddsaSignature);
-                var message = $"Sign this message to complete transaction\n\n{nftDetails.Name}\nQuantity: 1\nPrice: {takerOrderFee.Eth} LRC\nMaximum fees: {takerOrderFee.Fee} LRC\nSold by: {nftListingDetails.Maker}\nNFT data: {nftData}";
-                var l1Key = new EthECKey(settings.L1PrivateKey);
-                var signer = new EthereumMessageSigner();
-                var signedMessage = signer.EncodeUTF8AndSign(message, l1Key);
-
-                //Buy from LoopExchange
+                //Sign the order
+                (NftOrder nftTakerOrder, string takerEddsaSignature, string message, string signedMessage) = await Utils.CreateAndSignNftTakerOrderAsync(settings, nftDetails, nftTakerListingDetails, nftListingDetails, orderFee, storageId, takerOrderFee);
+                
+                //Submit the trade
                 var submitTrade = await loopExchangeWebApiService.SubmitTradeAsync(settings.LoopringAccountId, nftListingDetails.Id, nftTakerOrder, takerEddsaSignature, signedMessage);
                 if (submitTrade != null && submitTrade.ToString() == "{}")
                 {
-                    Console.WriteLine("Bought NFT successfully! Press any key to go back to the options");
+                    Console.WriteLine("Bought NFT successfully! Press any key to go back to the options!");
                     Console.ReadKey();
                     return true;
                 }
@@ -189,7 +135,6 @@ class Program
             return false;
         }
         return false;
-
     }
 }
